Allow specifying glob paths as package names
For example,

zplug '~/tmp/*.zsh', from:local
zplug '~/tmp/', from:local, use:'*.zsh'

are equivalent. If some of the paths do not exist, such as:

zplug '~/{a,b,some_nonexistent_file}.zsh', from:local

only the paths that exist are used, and if there is no such path, an error
message is displayed.