Fix fpath not updated in time for modules/prompt
With this change users don't have to specify:

    zplug 'modules/prompt', from:prezto, nice:10

but simply

    zplug 'modules/prompt', from:prezto

The init.zsh of modules/prompt has to be added as nice_plugins because
the fpath to modules/prompt/functions has to be added before promptinit,
which is called in init.zsh. This is due to how zplug loads plugins (in
the order of normal (nice < 9) plugins -> add fpath -> nice >= 10
plugins -> lazy plugins).