Add another install way
Add another install way which zsh will download and install the zplug automatically. Nothing will happen if it was already installed.
The benefit of this approach is that we don't need to run the install command manually. When restoring configurations on a new machine, the only thing to do is copy the zshrc file or make a symbol link.