Support git built from source
When git is built from source, the `git version` output contains
non-numeric version components.

```
$ git version
git version 2.24.0.rc2.46.g1d34d425d4
```

zplug cannot detect the git version with this output, and it fails to
start. Change the version pattern to extract a part that starts with a
digit. For the example above, zplug will detect the git version as
"2.24.0.rc2.46.g1d34d425d4". This is good enough for the version check
while there would be no false positives.